requirements.txt
bioblend
pyyaml
pytest-cov